requirements.txt
matplotlib==3.5.0
numpy==2.0.0
pillow